Add HLP to the Engine make process
