+++ src/engine/CMakeLists.txt
@@ -86,6 +86,7 @@ add_executable(main ${ENGINE_SOURCE_DIR}/main.cpp)
 add_subdirectory(${ENGINE_SOURCE_DIR}/catalog)
 add_subdirectory(${ENGINE_SOURCE_DIR}/server)
 add_subdirectory(${ENGINE_SOURCE_DIR}/router)
+add_subdirectory(${ENGINE_SOURCE_DIR}/hlp)
 # TODO dependent
 add_subdirectory(${ENGINE_SOURCE_DIR}/graph)
 add_subdirectory(${ENGINE_SOURCE_DIR}/json)
