+++ src/engine/source/hlp/LogQLParser.hpp
@@ -1,0 +1,9 @@
+#ifndef _LOGQL_PARSER_H
+#define _LOGQL_PARSER_H
+
+#include <string>
+#include "hlp.hpp"
+
+ParserList parseLogQlExpr(std::string const &expr);
+
+#endif //_LOGQL_PARSER_H
