+++ src/engine/test/CMakeLists.txt
@@ -25,3 +25,4 @@ add_subdirectory(${TEST_SOURCE_DIR}/graph)
 add_subdirectory(${TEST_SOURCE_DIR}/catalog)
 add_subdirectory(${TEST_SOURCE_DIR}/cliParser)
 add_subdirectory(${TEST_SOURCE_DIR}/shared)
+add_subdirectory(${TEST_SOURCE_DIR}/hlp)
