+++ src/engine/test/source/hlp/CMakeLists.txt
@@ -1,0 +1,23 @@
+
+CPMAddPackage(
+    NAME RapidJSON
+    GITHUB_REPOSITORY Tencent/rapidjson
+    GIT_TAG master
+    VERSION 1.1.0
+    OPTIONS
+        "RAPIDJSON_BUILD_DOC OFF"
+        "RAPIDJSON_BUILD_EXAMPLES OFF"
+        "RAPIDJSON_BUILD_TESTS OFF"
+)
+
+# Configuration Catalog tests
+add_executable(hlp_test
+  ${ENGINE_SOURCE_DIR}/hlp/hlp.cpp
+  ${ENGINE_SOURCE_DIR}/hlp/LogQLParser.cpp
+  ${ENGINE_SOURCE_DIR}/hlp/SpecificParsers.cpp
+  ${TEST_SOURCE_DIR}/hlp/hlp_test.cpp
+)
+
+target_include_directories(hlp_test PUBLIC "${ENGINE_SOURCE_DIR}/" "${RapidJSON_SOURCE_DIR}/include/")
+target_link_libraries(hlp_test gtest_main)
+gtest_discover_tests(hlp_test)
